remove `corda-avro-schema` from `uniqueness-common`
