+++ libs/uniqueness/common/build.gradle
@@ -10,9 +10,9 @@ dependencies {
 
     implementation 'org.jetbrains.kotlin:kotlin-osgi-bundle'
     implementation 'net.corda:corda-base'
-    implementation "net.corda:corda-avro-schema"
     implementation 'net.corda:corda-crypto'
     implementation 'net.corda:corda-ledger-utxo'
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     implementation project(':libs:crypto:crypto-core')
+    implementation libs.jackson.annotations
 }
