+++ libs/uniqueness/common/src/main/kotlin/net/corda/uniqueness/datamodel/internal/UniquenessCheckRequestInternal.kt
@@ -1,8 +1,5 @@
 package net.corda.uniqueness.datamodel.internal
 
-import net.corda.crypto.core.parseSecureHash
-import net.corda.data.uniqueness.UniquenessCheckRequestAvro
-import net.corda.uniqueness.datamodel.common.toStateRef
 import net.corda.v5.application.uniqueness.model.UniquenessCheckStateRef
 import net.corda.v5.crypto.SecureHash
 import java.time.Instant
@@ -21,39 +18,4 @@ data class UniquenessCheckRequestInternal constructor(
     val numOutputStates: Int,
     val timeWindowLowerBound: Instant?,
     val timeWindowUpperBound: Instant
-) {
-    companion object {
-        fun create(externalRequest: UniquenessCheckRequestAvro): UniquenessCheckRequestInternal {
-
-            with (externalRequest) {
-                require(numOutputStates >= 0) { "Number of output states cannot be less than 0." }
-
-                val duplicateInputs = inputStates.groupingBy { it }.eachCount().filter { it.value > 1 }
-
-                require(duplicateInputs.isEmpty()) { "Duplicate input states detected: ${duplicateInputs.keys}" }
-
-                val duplicateReferences = referenceStates.groupingBy { it }.eachCount().filter { it.value > 1 }
-
-                require(duplicateReferences.isEmpty()) { "Duplicate reference states detected: ${duplicateReferences.keys}" }
-
-                val intersection = inputStates intersect referenceStates.toSet()
-
-                require(intersection.isEmpty()) {
-                    "A state cannot be both an input and a reference input in the same request. Offending " +
-                            "states: $intersection"
-                }
-
-                return UniquenessCheckRequestInternal(
-                    parseSecureHash(txId),
-                    txId,
-                    originatorX500Name,
-                    inputStates?.map { it.toStateRef() } ?: emptyList(),
-                    referenceStates?.map { it.toStateRef() } ?: emptyList(),
-                    numOutputStates,
-                    timeWindowLowerBound,
-                    timeWindowUpperBound
-                )
-            }
-        }
-    }
-}
+)
